Audio: Change how volume works
Across the entire audio system, audio now works in 0-1 terms instead of
0-100 as before. Therefore, volume is now a double instead of an int.
The master volume of the AudioServer changes smoothly through a
FadingProperty, preventing clicks. Finally, volume computations are done
with logarithmic scaling, which is more natural for the human ear.

Note that this could be 4-5 different commits, but as they change each
other's code all the time, it makes no sense to split them up.

+++ Userland/Libraries/LibAudio/ClientConnection.cpp
@@ -48,7 +48,7 @@ void ClientConnection::muted_state_changed(bool muted)
         on_muted_state_change(muted);
 }
 
-void ClientConnection::main_mix_volume_changed(i32 volume)
+void ClientConnection::main_mix_volume_changed(double volume)
 {
     if (on_main_mix_volume_change)
         on_main_mix_volume_change(volume);
